+++ src/masks.cpp
@@ -42,12 +42,12 @@ void MasksParser::parse_allele(std::istream &stream) {
         if (a > allele_count)
             allele_count = a;
         if (a < allele_count && a != 0) {
-            MasksParser::allele_coverage.push_back(std::vector<int>(allele_count, 0));
+            MasksParser::allele_coverage.push_back(std::vector<float>(allele_count, 0));
             allele_count = a;
         }
         MasksParser::allele.push_back(a);
     }
 
     if (allele_count > 0)
-        MasksParser::allele_coverage.push_back(std::vector<int>(allele_count, 0));
+        MasksParser::allele_coverage.push_back(std::vector<float>(allele_count, 0));
 }
